requirements.txt
asgiref==3.8.1
Django==4.2
django-js-asset==2.2.0
django-mptt==0.16.0
Faker==24.11.0
install==1.3.5
isort==5.13.2
psycopg2==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
sqlparse==0.5.0
